Set up rudimentary project structure, this still needs a LOT of improvements
